board.py: Adjusted pieces to account for added variables (defenseMatrix, delegated) Fixed floodfill (now accurately calculates movement)
GameScene.py:
Fixed piece movement amounts.

main.py:
Imported board in order to remove highlights when changing a game state.

pieces.py:
Added two variables to the definition of a piece: defenseMatrix and delegated. DefenseMatrix is a tuple of (int, int) and allows the user to quickly reference the needed roll to claim a piece. Delegated allows us to quickly determine if a piece has been delegated by the king or not. I have plans of adding a black dot to delegated pieces to help identify them.