board.py: Removed tuples from piece creation.
GameFunctions.py:
Added attackMatrix and attack function. This function should be called anytime a piece is attempting to claim another piece. This function accepts the types of the attacker and defender, and references the attack matrix to see whether the attack is successful.

pieces.py:
Removed the tuple part of the pieces.