board.py: Comments for some methods
GameScene.py:
Debugging knight turn. Added extra checks for the commander's has_moved check. Additionally added in a check for if a Knight wants to move but it's not the one that has already triggered the special turn condition.

Additional checks for nearby enemies to help with the Knight that wants to move for adjacent enemies